Added singleEvent reducer, tested and working with allEvents page and singleEvent page
