+++ src/components/index.js
@@ -1,6 +1,7 @@
 export { default as userSlice } from "../store/users/userSlice";
 export { default as eventSlice } from "../store/events/eventSlice";
 export { default as cartSlice } from '../store/cart/cartSlice';
+export { default as singleEventSlice } from '../store/singleEvent/singleEventSlice';
 export { default as Navbar } from "./navbar/Navbar";
 export { default as EventList } from "./eventsList/EventList";
 export { default as LoginPage } from "./login/LoginPage";
@@ -13,4 +14,4 @@ export { default as EventTagListArts } from './eventsTagList/EventTagListArts'
 export { default as EventTagListMusic } from './eventsTagList/EventTagListMusic'
 export { default as EventTagListMisc } from './eventsTagList/EventTagListMisc'
 export { default as Checkout } from './checkout/Checkout'
-export {getSingleEvent} from '../store/events/eventSlice'
+export { getSingleEvent } from '../store/singleEvent/singleEventSlice'
