+++ src/components/singleEvent/SingleEvent.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useDispatch,useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { getSingleEvent } from "../";
 import { useParams } from 'react-router-dom'
