+++ src/store/events/eventSlice.js
@@ -27,16 +27,6 @@ export const getEventsByTag = createAsyncThunk('/allEventsByTag', async (tagId) 
     }
 })
 
-export const getSingleEvent = createAsyncThunk('/singleEvent', async (id) => {
-    try{
-        const { data } = await axios.get(`/api/events/${id}`)
-        return data
-    }
-    catch(err) {
-        console.log(err)
-    }
-})
-
 
 export const eventSlice = createSlice({
     name: 'events',
@@ -45,9 +35,6 @@ export const eventSlice = createSlice({
     extraReducers: {
         [getEvents.fulfilled]: (state, action) => {
             state.events = action.payload
-        },
-        [getSingleEvent.fulfilled]: (state, action) => {
-            return action.payload;
         },
         [getEventsByTag.fulfilled]: (state, action) => {
             state.events = action.payload;
