+++ src/store/index.js
@@ -1,13 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit'
 import reduxLogger from 'redux-logger';
-import { userSlice, eventSlice, cartSlice } from '../components'
+import { userSlice, eventSlice, cartSlice, singleEventSlice } from '../components'
 
 
 const store = configureStore({
     reducer: {
         users: userSlice,
         events: eventSlice,
-        cart: cartSlice
+        cart: cartSlice,
+        singleEvent: singleEventSlice
     },
     middleware: (getDefaultMiddleware) =>
       getDefaultMiddleware().concat(reduxLogger)
